requirements.txt
# file: requirements.txt

# --- Core Framework ---
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# --- Vector DB & Embeddings ---
qdrant-client
fastembed

# --- UI & Utilities ---
gradio
python-dotenv
pyyaml
httpx

# --- Log Generation ---
faker